iterate on IntUiBridgeTab
